// src/models/client/client-store-model.ts
export interface ClientStore {
  name: string;
  email: string;
  phoneNumber: string;
  address: string;
  postalCode: string;
  facebookUrl: string;
  instagramUrl: string;
  linkedInUrl: string;
  websiteUrl: string;
  whatsapp: string;
  emailContact: string;
  about: string;
  logo: string;
  currencySymbol: string;
  countryName: string;
  city: string;
}